preview: extract shared static-site AWS helpers to aws_utils; move preview deploy to agenttools
The static preview deploy core doesn't belong in jobs/commands — Commands
implement Run and are dispatched sequentially by the job engine, whereas the
preview deploy is called in-process by the agent_mcp tool handlers. Move it to
a new top-level agenttools package (the home for agent-invoked MCP tool
implementations).

Extract the S3 + CloudFront primitives shared by the production static deploy
and the preview deploy (CreateS3BucketIfNeeded, UploadToS3, CreateCachePolicy,
CreateOriginAccessControl, CreateDefaultCacheBehavior, DeleteAllS3Files,
AttachPolicyToS3Bucket) into the leaf utils/aws_utils package as exported
functions, so agenttools can reuse them without importing jobs/commands (which
would be an import cycle). deploy_aws_static_site and delete_aws_static_site now
call aws_utils.* for these primitives.

No behavior change — pure move + rename.
